[#8] feat: 삭제 버튼 클릭시 확인 콘솔을 띄우는 기능 추가
삭제 버튼을 클릭시 '정말 지우시겠습니까?'라는 메세지와 함게
확인 콘솔을 띄우는 기능을 추가하였습니다.

+++ src/containers/CommentListContainer.js
@@ -23,6 +23,9 @@ function CommentListContainer() {
     dispatch(getForm(id));
   };
   const createDeleteHandler = id => () => {
+    const agree = confirm('정말 삭제하시겠습니까?');
+    if (!agree) return;
+
     commentApi.deleteAComment(id);
     dispatch(deleteComments(id));
     dispatch(movePage(1));
